+++ CHANGELOG.md
@@ -10,6 +10,7 @@ v3.23.1 - 2018-09-??
 + Updated all access plan templates added in 3.23.0 to ensure `ABSPATH` is defined to prevent direct template access
 + Refactored the `LLMS_Section->get_percent_complete()` method to utilize methods from the `LLMS_Student` model
 + Added the ability for admin table classes to define `<tr>` element CSS classes
++ Admin settings pages with no settings to save (like the Notifications list) no longer display a "Save" button
 
 ##### Deprecated Functions & Methods